// lib/openapi/models/CreateSubscriptionSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateSubscriptionSchema = {
    stationId: string;
    url: string;
    messageRegexFilter?: string;
    onClose?: boolean;
    onConnect?: boolean;
    onMessage?: boolean;
    sentMessage?: boolean;
};